+++ Sources/KeyboardKit/Autocomplete/AutocompleteToolbarStyle.swift
@@ -20,13 +20,13 @@ public struct AutocompleteToolbarStyle {
      Create an autocomplete toolbar style.
      
      - Parameters:
-       - height: The toolbar height, by default `50`.
+       - height: An optional toolbar height, by default `50`.
        - item: The item style, by default `.standard`.
        - separator: The separator style, by default `.standard`.
        - autocompleteBackground: The background style to apply behind autocomplete items, by default `.standard`.
      */
     public init(
-        height: Double = 50,
+        height: CGFloat? = 50,
         item: AutocompleteToolbarItemStyle = .standard,
         separator: AutocompleteToolbarSeparatorStyle = .standard,
         autocompleteBackground: AutocompleteToolbarItemBackgroundStyle = .standard
@@ -43,9 +43,11 @@ public struct AutocompleteToolbarStyle {
     public var autocompleteBackground: AutocompleteToolbarItemBackgroundStyle
 
     /**
-     The toolbar height.
+     An optional toolbar height.
+
+     Set it to nil to avoid applying a strict toolbar height.
      */
-    public var height: Double
+    public var height: CGFloat?
     
     /**
      The item style to apply to the toolbar items.