+++ RELEASE_NOTES.md
@@ -24,6 +24,7 @@ Some things that are planned for the 7.0 release:
 
 ### ✨ New features
 
+* `AutocompleteToolbarStyle` has a new optional `height` parameter.
 * `CharacterProvider` is a new protocol that will let us design better character-based extensions over time.
 * `EmojiCharacterAnalyzer` is a new protocol that is implemented by `Character`.
 * `EmojiStringAnalyzer` is a new protocol that is implemented by `String`.
@@ -31,6 +32,11 @@ Some things that are planned for the 7.0 release:
 * `KeyboardRootView` no longer defines an explicit id.
 * `LocaleContextMenu` is a new view modifier for applying a locale context menu to any view.
 * `StringProvider` is a new protocol that will let us design better string-based extensions over time.
+
+### 💡 Behavior changes
+
+* `AutocompleteToolbar` now applies the height from its style.
+* KeyboardKit no longer applies an explicit id to its content view.
 
 ### 🗑 Deprecations